Allow Hotglue to work behind a load balancer or a reverse proxy
In modern servers environment, we frequently put a front-end server
to offer a cache (e.g. Varnish), load balancing or allow to isolate
application in Docker containers.

In such configurations, the responsibility to serve in TLS mode
belongs to the front-end. We need to play along and serve https://
URL, not http:// ones in such case.

The most used header (Apache, Varnish, nginx) to indicate this state
is X-Forwarded-Proto, normalised as the Forwarded HTTP Extension.

This is coherent with RFC 7239 5.4.
https://tools.ietf.org/html/rfc7239#section-5.4