Get rid of array_shift
The array_shift PHP function expects an array to manipulate,
stored in a variable. As such, it's a poor candidate to get
the first element of an array, as it will raise a strict
exception if we pass it a reference to an array instead.

We supersede array_shift by an universal function to get
the first element of any array or object which implements
the Traversable interface, kicking the iterator to pick
th first element.

Fixes #9.